database/scripts/requirements.txt
osmium
requests
psycopg2-binary
python-dotenv